requirements.txt
Django>=4.1.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
psycopg2>=2.9.0,<2.9.5
uWSGI>=2.0.19.1,<2.1
drf-spectacular>=0.26.0,<1.0.0
Pillow>=9.4.0,<10.0.0